+++ CHANGES.rst
@@ -7,6 +7,7 @@ Changelog
 - Fix upload with int folderid #63 [tomgross]
 - Add pytest timeout and update testing dependencies [tomgross]
 - Implement `copyfile` and `downloadfileasync` methods [tomgross]
+- Implement `setlanguage`, `getfeedback`, `diff` & `getfilehistory` methods [tomgross]
 
 
 1.0 (2022-02-02)